+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - path as query parameter in order to remember path in history
+- integration with Keycloak
 
 ### Removed